pivot from family-specific KL divergences towards PyTorch pointwise KL divergence loss method for simplicity
