+++ src/kulprit/projection/solvers/gradient.py
@@ -7,7 +7,7 @@ import torch
 from kulprit.data.data import ModelData
 from kulprit.families.family import Family
 from kulprit.projection.architecture import Architecture
-from kulprit.projection.losses.kld import KullbackLeiblerLoss
+from kulprit.projection.losses import KullbackLeiblerLoss
 from kulprit.projection.solvers import BaseSolver
 
 
@@ -43,7 +43,7 @@ class GradientDescentSolver(BaseSolver):
 
         # build architecture and loss methods for gradient descent
         self.architecture = Architecture(submodel_structure)
-        self.loss = KullbackLeiblerLoss(self.family)
+        self.loss = KullbackLeiblerLoss()
 
         # extract submodel design matrix
         X_perp = submodel_structure.X
@@ -76,14 +76,13 @@ class GradientDescentSolver(BaseSolver):
         # build optimisation framework
         solver = self.architecture.architecture
         solver.zero_grad()
-        loss_fun = self.loss.loss
         optim = torch.optim.Adam(solver.parameters(), lr=self.learning_rate)
 
         # run optimisation loop
         for _ in range(self.num_iters):
             optim.zero_grad()
             y_perp = solver(X_perp)
-            loss = loss_fun.forward(y_ast, y_perp)
+            loss = self.loss.forward(y_perp, y_ast)
             loss.backward()
             optim.step()
